Add Atmel pubkeys
Atmel only provide public keys rather than X509 certificates, so we need
to handle this slightly differently. We assume that the EK certificate is
directly signed with the private component of the key, so construct a key
from the data they provided us (we're guessing that the exponent is 65537)
and verify the EK certificate with that.